% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_block.R
\name{fd_block_fn_label}
\alias{fd_block_fn_label}
\title{Create a readable label for a function block}
\usage{
fd_block_fn_label(fn)
}
\arguments{
\item{fn}{A function.}
}
\value{
Character scalar label.
}
\description{
Create a readable label for a function block
}
\keyword{internal}
