% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_compute.R
\name{fd_compute_call_block_function}
\alias{fd_compute_call_block_function}
\title{Call a block function against the current working data}
\usage{
fd_compute_call_block_function(fn, working_data)
}
\arguments{
\item{fn}{Function stored in a function-mode block.}

\item{working_data}{Current working data.}
}
\value{
Block result.
}
\description{
Call a block function against the current working data
}
\keyword{internal}
