% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_compute.R
\name{fd_compute_eval_column_step}
\alias{fd_compute_eval_column_step}
\title{Evaluate one column step}
\usage{
fd_compute_eval_column_step(
  step,
  step_name,
  mask,
  n,
  compute_envir = NULL,
  compute_strict = TRUE
)
}
\arguments{
\item{step}{Column step.}

\item{step_name}{Step name.}

\item{mask}{Current evaluation mask.}

\item{n}{Expected row count.}

\item{compute_envir}{Optional overriding environment.}

\item{compute_strict}{Logical strictness flag.}
}
\value{
A list with elements \code{ok}, \code{error}, and \code{cols}.
}
\description{
Evaluate one column step
}
\keyword{internal}
