% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_define.R
\name{normalize_block_step}
\alias{normalize_block_step}
\title{Normalize a block definition into a step object}
\usage{
normalize_block_step(x, env, step_name)
}
\arguments{
\item{x}{A block definition object.}

\item{env}{Environment used for evaluation.}

\item{step_name}{Step name used for clearer error messages.}
}
\value{
A normalized block step object.
}
\description{
Internal placeholder for future \code{fd_block()} support.
}
\seealso{
Other featdelta defs helpers: 
\code{\link{fd_block}()},
\code{\link{fd_define}()},
\code{\link{fd_define_validate_step_defs}()},
\code{\link{fd_define_validate_step_names}()},
\code{\link{maybe_eval_fd_block_call}()},
\code{\link{maybe_unwrap_symbol}()},
\code{\link{normalize_definition_step}()},
\code{\link{normalize_feature_input}()},
\code{\link{print.featdelta_defs}()},
\code{\link{summary.featdelta_defs}()}
}
\concept{featdelta defs helpers}
\keyword{internal}
