% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_run.R
\name{print.fd_run_report}
\alias{print.fd_run_report}
\title{Print a featdelta run report}
\usage{
\method{print}{fd_run_report}(x, ...)
}
\arguments{
\item{x}{An \code{fd_run_report} object.}

\item{...}{Unused.}
}
\value{
The input object invisibly.
}
\description{
Print a featdelta run report
}
