% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_connect.R
\name{print.featdelta_con}
\alias{print.featdelta_con}
\title{Print a featdelta connection context}
\usage{
\method{print}{featdelta_con}(x, ...)
}
\arguments{
\item{x}{A \code{featdelta_con} object.}

\item{...}{Unused.}
}
\value{
The input object invisibly.
}
\description{
Print a featdelta connection context
}
